+++ covalent_data/app/core/s3storagebackend.py
@@ -49,7 +49,12 @@ class S3StorageBackend(ABC):
         err_code = e.response["Error"]["Code"]
         logger.warn(f"S3 Client Error: {err_code}")
         if err_code == "AccessDenied":
-            raise HTTPException(401, detail="AccessDenied: Likely incorrect AWS credentials.")
+            raise HTTPException(
+                401,
+                detail="S3 client error: AccessDenied (Likely incorrect AWS credentials or configuration).",
+            )
+        else:
+            raise HTTPException(400, detail=f"S3 client error: {err_code}")
 
     def get(self, bucket_name: str, object_name: str) -> Union[Generator[bytes, None, None], None]:
         """Get object from storage.